+++ src/Streamly/Data/Stream/Concurrent.hs
@@ -1,0 +1,14 @@
+-- |
+-- Module      : Streamly.Data.Stream.Concurrent
+-- Copyright   : (c) 2022 Composewell Technologies
+-- License     : BSD-3-Clause
+-- Maintainer  : streamly@composewell.com
+-- Stability   : released
+-- Portability : GHC
+
+module Streamly.Data.Stream.Concurrent
+  ( MonadAsync
+  )
+where
+
+import Streamly.Internal.Control.Concurrent (MonadAsync)